+++ CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New command to show Sedge container info: `sedge show`
+- Add default [checkpoint sync url]( https://checkpoint.chiadochain.net ) for Chiado.
+- New `--skip-pull` flag to skip pulling docker images when running `sedge run`.
+
+### Fixed
+
+- Change validator blocker container image to [busybox](https://hub.docker.com/_/busybox).
 
 ## [v1.1.0] - 2023-04-07